+++ CitySimulator/src/core/include/core/data_layer/way_info.h
@@ -23,6 +23,7 @@ namespace tjs::core {
 		Bus_Stop, Emergency_Access, Delivery_Access);
 
 	struct Node;
+	struct Coordinates;
 
 	ENUM_FLAG(TurnDirection, char,
 		None = 0,
@@ -31,8 +32,8 @@ namespace tjs::core {
 		Straight = 1 << 2,
 		UTurn = 1 << 3,
 		MergeRight = 1 << 4,
-		MergeLeft = 1 << 5
-	);
+		MergeLeft = 1 << 5);
+	TurnDirection get_relative_direction(const Coordinates& a, const Coordinates& o, const Coordinates& b, bool rhs = true);
 
 	struct WayInfo {
 		uint64_t uid;
